+++ code/ryzom/server/src/ai_service/ai_bot_npc.cpp
@@ -496,7 +496,7 @@ void CBotNpc::calcSpawnPos(RYAI_MAP_CRUNCH::CWorldMap const& worldMap)
 		std::vector<CShape::TPosition> const& posList = state->shape().getGeometry();		
 		do
 		{
-			const	uint32 a=CAIS::rand16(posList.size());
+			const	uint32 a=CAIS::rand16((uint32)posList.size());
 			const	uint32 b=(a+1)%posList.size();
 			const	double weight=CAIS::frand();
 			_StartPos.setXY(posList[a].toAIVector()+(posList[b].toAIVector()-posList[a].toAIVector())*weight);
@@ -973,7 +973,7 @@ void CBotNpc::setColours(std::string input)
 						if (NLMISC::nlstricmp(colour,colourNames[i][j])==0)
 						{
 							// found an entry so add to the results vector
-							results.push_back(i);
+							results.push_back((uint32)i);
 							done=true;
 						}
 					}
@@ -1003,7 +1003,7 @@ void CBotNpc::setColours(std::string input)
 		{
 			// upper body colour
  			generator.srand(seed+975*0x10000);
-			uint8 val = (uint8)results[generator.rand(results.size()-1)];
+			uint8 val = (uint8)results[generator.rand((uint16)results.size()-1)];
 			_Sheet->setColorArms(val);
 			_Sheet->setColorHands(val);
 			_Sheet->setColorBody(val);
@@ -1012,7 +1012,7 @@ void CBotNpc::setColours(std::string input)
 		{
 			// lower body colour
  			generator.srand(seed+977*0x10000);
-			uint8 val = (uint8)results[generator.rand(results.size()-1)];
+			uint8 val = (uint8)results[generator.rand((uint16)results.size()-1)];
 			_Sheet->setColorLegs(val);
 			_Sheet->setColorFeets(val);
 		}
@@ -1020,49 +1020,49 @@ void CBotNpc::setColours(std::string input)
 		{
 			// hair colour, mapped to head
  			generator.srand(seed+976*0x10000);
-			uint8 val = (uint8)results[generator.rand(results.size()-1)];
+			uint8 val = (uint8)results[generator.rand((uint16)results.size()-1)];
 			_Sheet->setColorHead(val);
 		}
 		else if ( NLMISC::nlstricmp(keyword,"CHEAD")==0)
 		{
 			// head color
  			generator.srand(seed+979*0x10000);
-			uint8 val = (uint8)results[generator.rand(results.size()-1)];
+			uint8 val = (uint8)results[generator.rand((uint16)results.size()-1)];
 			_Sheet->setColorHead(val);
 		}
 		else if ( NLMISC::nlstricmp(keyword,"CARMS")==0)
 		{
 			// arms color
  			generator.srand(seed+981*0x10000);
-			uint8 val = (uint8)results[generator.rand(results.size()-1)];
+			uint8 val = (uint8)results[generator.rand((uint16)results.size()-1)];
 			_Sheet->setColorArms(val);
 		}
 		else if ( NLMISC::nlstricmp(keyword,"CHANDS")==0)
 		{
 			// arms color
  			generator.srand(seed+983*0x10000);
-			uint8 val = (uint8)results[generator.rand(results.size()-1)];
+			uint8 val = (uint8)results[generator.rand((uint16)results.size()-1)];
 			_Sheet->setColorHands(val);
 		}
 		else if ( NLMISC::nlstricmp(keyword,"CBODY")==0)
 		{
 			// arms color
  			generator.srand(seed+985*0x10000);
-			uint8 val = (uint8)results[generator.rand(results.size()-1)];
+			uint8 val = (uint8)results[generator.rand((uint16)results.size()-1)];
 			_Sheet->setColorBody(val);
 		}
 		else if ( NLMISC::nlstricmp(keyword,"CLEGS")==0)
 		{
 			// arms color
  			generator.srand(seed+987*0x10000);
-			uint8 val = (uint8)results[generator.rand(results.size()-1)];
+			uint8 val = (uint8)results[generator.rand((uint16)results.size()-1)];
 			_Sheet->setColorLegs(val);
 		}
 		else if ( NLMISC::nlstricmp(keyword,"CFEETS")==0)
 		{
 			// arms color
  			generator.srand(seed+989*0x10000);
-			uint8 val = (uint8)results[generator.rand(results.size()-1)];
+			uint8 val = (uint8)results[generator.rand((uint16)results.size()-1)];
 			_Sheet->setColorFeets(val);
 		}
 	}
