+++ code/ryzom/server/src/ai_service/ai_entity_physical.h
@@ -305,7 +305,7 @@ public:
 	//@{
 	virtual void addHealer(IAIEntityPhysicalHealer* healer) { _Healers.insert(healer); if (healer) healer->healerAdded(this); }
 	virtual void delHealer(IAIEntityPhysicalHealer* healer) { _Healers.erase(healer); if (healer) healer->healerRemoved(this); }
-	virtual int getHealerCount() { return _Healers.size(); }
+	virtual int getHealerCount() { return (int)_Healers.size(); }
 	//@}
 	
 	static int _PlayerVisibilityDistance;
