+++ code/ryzom/server/src/ai_service/ai_script_comp.cpp
@@ -989,7 +989,7 @@ public:
 		vector<string>	params;
 		explodeSubStrings(inStr, params, -1);
 		
-		const	uint32	nbSubScript=params.size();
+		const	uint32	nbSubScript=(uint32)params.size();
 		
 		std::vector<CSmartPtr<CFightScriptComp> >	scriptComps;
 		try
